Migrate content-downloader from js to typescript
