+++ content-downloader/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 eslint_report.json
+dist
