--- content-downloader/bin/hugo-content-downloader
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require = require('esm')(module /*, options*/);
-require('../src/cli').cli(process.argv);
