+++ content-downloader/sonar-project.properties
@@ -8,6 +8,8 @@ sonar.projectName=website-favre: content-downloader
 sonar.sources=.
 
 # Encoding of the source code. Default is default system encoding
-sonar.sourceEncoding=UTF-8
-
+sonar.sourceEncoding = UTF-8
+sonar.language = ts
+sonar.exclusions = **/node_modules/**, **/dist/**
+sonar.typescript.tsconfigPath=  ./tsconfig.json
 sonar.eslint.reportPaths = ./eslint_report.json
