+++ content-downloader/src/svg.ts
